Test to ensure that the output is unchanged.
This is important to ensure that we do not introduce more bugs while
refactoring. We should have more detailed tests to check that the program's
working as designed, but I don't know enough about what you're trying to do to
write those.

You can run this test by typing "prove" (bundled with perl) in the repository's
top-level directory. It's a good idea to get into the habit of running your
tests as often as possible.