More input data checking; fix broken data.
Sanity-checking your input data is a good idea anyway, but in this case the
input data file contained the non-numeric entry "1066.3kj" on the last line.
We can check for malformed data like this using the looks_like_number function
in the core module Scalar::Util. The "kj" looks like vi commands, so it's
probably my fault that it got in there. But anyway! Now our program is more
robust.

To my surprise, the program's output didn't change at all.