Replace pairs of arrays with arrays of pairs.
We replace @nameArray and @dataArray with an array @genes, whose elements are
anonymous hashes containing a name and an arrayref of data. Similarly, we
replace @filterData and @filterNames with a single array @filtered. This
increases memory consumption by a bit less than 10% (6.5MB versus 6MB,
according to Devel::Size), but simplifies the subsequent logic and allows us to
eliminate the $indexLine loop counter variable. We can't eliminate the counter
in the second loop, because we need it to report the current cycle.

But mainly, having related bits of data that were only tied together by
convention ("$name[$i] is the name of the gene sampled in $data[$i]") made me
nervous. Much better and more robust to let the interpreter enforce that
relationship for you.