Replace for-loop with call to grep.
The grep builtin, somewhat like its Unix namesake, takes an array @x and a
predicate and returns an array of all those elements of @x for which the
predicate is true. You can read its full documentation by typing `perldoc -f
grep` at the command line.

In this case the grep call's a bit noisy and punctuation-heavy: the loop it
replaced might have been clearer. TMTOWTDI.