Replace explicit open/close with magic <> operator.
<> allows you to easily implement classic Unix filters: read all the files
specified on the command-line in order, or stdin if no arguments are given. The
name of the current file is held in $ARGV.

I'm not sure what the correct behaviour should be for multiple data files, so
I've left the `@ARGV != 1` check in. If you want to handle multiple files with
<>, you'll need to alter the while loop a bit so that errors are reported with
the correct line number: see `perldoc -f eof` for details.

[Hat-tip to Aaron Crane]