Replace hand-rolled stats code with Statistics::Lite.
One of the most important habits to get into as a Perl programmer is the habit
of asking yourself "is there a CPAN module that could help me out here?" Hint:
yes, there is. In this case, Statistics::Lite allows you to (among other
things) calculate means and standard deviations without writing
potentially-buggy code yourself. Another possibility might have been
Statistics::Basic, but that only provides code to calculate the standard
deviation of the sample and not the sample standard deviation (see
http://en.wikipedia.org/wiki/Standard_deviation#With_standard_deviation_of_the_sample).

The best way to search for relevant CPAN modules is currently
http://metacpan.org , which also allows you to browse reverse dependencies,
per-platform test results and so on.

While I was making this change, I took the opportunity to delete the variables
$controlMean, $controlStdev etc, which were used exactly once.