Move filtering & calculation into while loop
This will reduce memory usage significantly, since we no longer store the
arrays @genes and @filtered. Not worth it for a few thousand lines, but
definitely worth it if you're processing multi-gigabyte files. Of course, you
probably don't want to do that with the script as it is now, since %score still
holds all the data in memory! You could perhaps maintain a priority queue of the
top N genes (see the CPAN module List::Priority for this) or use some sort of
tied hash if you really need them all.

We also factor out the calculation of the FLD score into a subroutine. This is,
I think, a win from the point of view of clarity.

The output's slightly changed; we used to know how many genes met the filter
criteria before we started calculating FLD scores, but we don't any more. I've
removed those lines from the expected output.